Implement ETT for selected pages
If a page has type user-guide-page, it loads the ETT theme, otherwise it loads the classic theme. That meant renaming existing layout files that had the same names as ETT files. Implemented for the registering.md page only so far.

The config settings for adding the userguide sidebar across all pages in the help/user-guide path is not working.

The theme colour was stuck for me with the default, but I have issues with caching of theme colour in other repos too.